Use clear int32_t representation for the new double to int function.
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@10390 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_util.cpp
@@ -2370,7 +2370,7 @@ int32_t util_t::magicnumber_doubletoint( double d )
   union Cast
   {
      double d;
-     long l;
+     int32_t l;
    };
   volatile Cast c;
   c.d = d + 6755399441055744.0;
